backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
firebase-admin
httpx
pandas
scikit-learn
prophet
numpy
python-jose[cryptography]
pydantic-settings